long metric

+++ src/libmetric/includes/metric.h
@@ -4,8 +4,8 @@
 #include <time.h>
 
 typedef struct metric_t {
-  int moves;
-  int comparisons;
+  long int moves;
+  long int comparisons;
   clock_t time;
 } Metric;
 
@@ -13,8 +13,8 @@ void metricInit(Metric *metric);
 int metricGetMoves(Metric *metric);
 int metricGetComparisons(Metric *metric);
 double metricGetSeconds(Metric metric);
-void metricSetMoves(Metric *metric, int moves);
-void metricSetComparisons(Metric *metric, int comparisons);
+void metricSetMoves(Metric *metric, long int moves);
+void metricSetComparisons(Metric *metric, long int comparisons);
 void metricFinish(Metric *metric);
 
 #endif
